backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.28.1
sqlalchemy[asyncio]==2.0.36
aiomysql==0.2.0
alembic==1.14.1
authlib==1.3.2
itsdangerous==2.2.0
pytest==8.3.4
pytest-asyncio==0.25.3
anyio==4.8.0
boto3==1.35.99
sentry-sdk[fastapi]>=2.0,<3.0
